requirements.txt
absl-py==0.12.0
astor==0.8.1
attrs==20.3.0
blis==0.7.4
cached-property==1.5.2
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
contextvars==2.4
cryptography==3.4.7
cymem==2.0.5
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
dill==0.3.3
dlx==1.0.4
docker-pycreds==0.4.0
docplex==2.15.194
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
fastdtw==0.3.4
fastjsonschema==2.15.0
filelock==3.0.12
gast==0.2.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.36.1
h5py==3.1.0
idna==2.10
immutables==0.15
importlib-metadata==3.8.0
inflection==0.5.1
isodate==0.6.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
more-itertools==8.7.0
mpmath==1.2.1
multitasking==0.0.9
murmurhash==1.0.5
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
ntlm-auth==1.5.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pathtools==0.1.2
pathy==0.4.0
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.1.2
pkg-resources==0.0.0
ply==3.11
plyfile==0.7.3
preprocessor==1.1.3
preshed==3.0.5
promise==2.3
protobuf==3.15.6
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyparsing==2.4.7
pyrouge==0.1.3
pyrsistent==0.17.3
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
qiskit==0.24.1
qiskit-aer==0.7.6
qiskit-aqua==0.8.2
qiskit-ibmq-provider==0.12.2
qiskit-ignis==0.5.2
qiskit-terra==0.16.4
Quandl==3.6.1
rdflib==5.0.0
regex==2021.3.17
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retworkx==0.8.0
rsa==4.7.2
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.5.4
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
smart-open==3.0.0
smmap==4.0.0
sortedcontainers==2.4.0
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
subprocess32==3.5.4
sympy==1.7.1
tensorboard==2.0.2
tensorboardX==2.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.3.0+cu100
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.4.0
torch-sparse==0.4.3
torchtext==0.4.0
torchvision==0.4.1+cu100
tqdm==4.59.0
transformers==4.4.2
tweet-preprocessor==0.6.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Wand==0.6.6
wandb==0.10.23
wasabi==0.8.2
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.59
zipp==3.4.1
